feat: arigmetic resolver for scanner

+++ wyrm_core/utils/token_vector.h
@@ -5,14 +5,15 @@ Como vamos a manjear un vector de toknes, usamos una estrucura auxilar,
 sombre todo para no lidar con problemas de dimencionamiento y memoria.
 */
 
+#include <stdlib.h>
 #include "../wyrm_token.h"
 
 
 const int ERROR_TOKEN_VECTOR = -1;
 
 typedef struct {
-    int size;
-    int capacity;
+    size_t size;
+    size_t capacity;
     wyrm_token_t* tokens;
 } TokenVector;
 
@@ -20,7 +21,7 @@ typedef TokenVector token_vector_t;
 
 
 // Crea un vector de tokens con la capacidad especificada
-token_vector_t* create_token_vector(int capacity);
+token_vector_t* create_token_vector(size_t capacity);
 
 // Agrega un token al vector, devuelve ERROR_TOKEN_VECTOR en caso de error
 int push_token(token_vector_t* vector, wyrm_token_t token);
